entity service controller for comment

+++ src/post/post.service.ts
@@ -22,10 +22,13 @@ export class PostService {
     const skip = items_per_page * (page - 1);
     const [res, total] = await this.postRepository.findAndCount({
       order: { updated_at: 'DESC' },
-      relations: ['user'],
+      relations: ['user', 'comments'],
       where: {
         title: Like(`%${search}%`),
-        user: { first_name: Like(`%${search}%`),last_name: Like(`%${search}%`) }
+        user: {
+          first_name: Like(`%${search}%`),
+          last_name: Like(`%${search}%`),
+        },
       },
       take: items_per_page,
       skip: skip,
@@ -54,10 +57,10 @@ export class PostService {
     };
   }
 
-  async findOne(id: string): Promise<Post> {
+  async findOne(id:number): Promise<Post> {
     return this.postRepository.findOne({
       where: { id },
-      relations: ['user'],
+      relations: ['user','comments'],
       select: {
         user: {
           id: true,
@@ -70,7 +73,7 @@ export class PostService {
     });
   }
 
-  async create(userId: string, createPostDto: CreatePostDto): Promise<Post> {
+  async create(userId: number, createPostDto: CreatePostDto): Promise<Post> {
     const user = await this.userRepository.findOneBy({ id: userId });
     try {
       const res = await this.postRepository.save({
@@ -85,8 +88,8 @@ export class PostService {
   }
 
   async update(
-    id: string,
-    userId: string,
+    id: number,
+    userId: number,
     postData: Partial<Post>,
   ): Promise<UpdateResult> {
     const post = await this.postRepository.findOne({
@@ -110,8 +113,26 @@ export class PostService {
     }
     return await this.postRepository.update({ id }, postData);
   }
-
-  async remove(id: string): Promise<void> {
+  async remove(id: number, userId: number): Promise<void> {
+    const post = await this.postRepository.findOne({
+      where: { id },
+      relations: ['user'],
+      select: {
+        user: {
+          id: true,
+          first_name: true,
+          last_name: true,
+          email: true,
+          avatar: true,
+        },
+      },
+    });
+    if (post.user.id !== userId) {
+      throw new HttpException(
+        'You are not authorized to delete this post',
+        HttpStatus.UNAUTHORIZED,
+      );
+    }
     await this.postRepository.delete(id);
   }
 }
